+++ tests/commands/mod.rs
@@ -2,5 +2,8 @@ mod exporter;
 mod exporter_links;
 mod exporter_option_git_ref;
 mod exporter_option_on_changes;
+mod exporter_option_remove_page_suffix;
 mod exporter_options;
+mod exporter_path;
 mod exporter_types;
+mod utils;